Separated functions to different folers, add README
